+++ sql/functions/create.function.neighborsInCats.sql
@@ -5,7 +5,7 @@ CREATE FUNCTION neighborsInCats(ira DOUBLE
                                ) RETURNS TABLE (id INT
                                                ,name VARCHAR(50)
                                                ,catsrcname VARCHAR(120)
-                                               ,band INT
+                                               ,band SMALLINT
                                                ,freq_eff DOUBLE
                                                ,ra DOUBLE
                                                ,decl DOUBLE
